fixed installChromeFromRepo

+++ bash/system_reinstall.sh
@@ -555,10 +555,12 @@ configurePhotosSync() {
 }
 
 installChromeFromRepo() {
-    wget -q -O - https://dl-ssl.google.com/linux/linux_signing_key.pub | sudo apt-key add -
-    sudo sh -c 'echo "deb [arch=amd64] http://dl.google.com/linux/chrome/deb/ stable main" >> /etc/apt/sources.list.d/google.list'
-    sudo apt-get update
-    aptInstall google-chrome-stable
+    if ! commandExists google-chrome; then
+        wget -q -O - https://dl-ssl.google.com/linux/linux_signing_key.pub | sudo apt-key add -
+        sudo sh -c 'echo "deb [arch=amd64] http://dl.google.com/linux/chrome/deb/ stable main" >> /etc/apt/sources.list.d/google.list'
+        sudo apt-get update
+        aptInstall google-chrome-stable
+    fi
 }
 
 main
